+++ src/gui/main_window.cpp
@@ -1,0 +1,215 @@
+/// @file main_window.cpp
+/// @brief Main application window implementation
+
+#include "kalahari/gui/main_window.h"
+#include "kalahari/core/logger.h"
+#include <QMenuBar>
+#include <QToolBar>
+#include <QStatusBar>
+#include <QApplication>
+
+namespace kalahari {
+namespace gui {
+
+MainWindow::MainWindow(QWidget* parent)
+    : QMainWindow(parent)
+    , m_newAction(nullptr)
+    , m_openAction(nullptr)
+    , m_saveAction(nullptr)
+    , m_saveAsAction(nullptr)
+    , m_exitAction(nullptr)
+    , m_undoAction(nullptr)
+    , m_redoAction(nullptr)
+    , m_cutAction(nullptr)
+    , m_copyAction(nullptr)
+    , m_pasteAction(nullptr)
+    , m_fileMenu(nullptr)
+    , m_editMenu(nullptr)
+    , m_fileToolbar(nullptr)
+{
+    auto& logger = core::Logger::getInstance();
+    logger.info("MainWindow constructor called");
+
+    // Set window properties
+    setWindowTitle("Kalahari Writer's IDE");
+    resize(1280, 720);
+
+    // Create UI components
+    createActions();
+    createMenus();
+    createToolbars();
+    createStatusBar();
+
+    logger.info("MainWindow initialized successfully");
+}
+
+void MainWindow::createActions() {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Creating actions");
+
+    // File actions
+    m_newAction = new QAction(tr("&New"), this);
+    m_newAction->setShortcut(QKeySequence::New);
+    m_newAction->setStatusTip(tr("Create a new document"));
+    connect(m_newAction, &QAction::triggered, this, &MainWindow::onNewDocument);
+
+    m_openAction = new QAction(tr("&Open..."), this);
+    m_openAction->setShortcut(QKeySequence::Open);
+    m_openAction->setStatusTip(tr("Open an existing document"));
+    connect(m_openAction, &QAction::triggered, this, &MainWindow::onOpenDocument);
+
+    m_saveAction = new QAction(tr("&Save"), this);
+    m_saveAction->setShortcut(QKeySequence::Save);
+    m_saveAction->setStatusTip(tr("Save the current document"));
+    connect(m_saveAction, &QAction::triggered, this, &MainWindow::onSaveDocument);
+
+    m_saveAsAction = new QAction(tr("Save &As..."), this);
+    m_saveAsAction->setShortcut(QKeySequence::SaveAs);
+    m_saveAsAction->setStatusTip(tr("Save the current document with a new name"));
+    connect(m_saveAsAction, &QAction::triggered, this, &MainWindow::onSaveAsDocument);
+
+    m_exitAction = new QAction(tr("E&xit"), this);
+    m_exitAction->setShortcut(QKeySequence::Quit);
+    m_exitAction->setStatusTip(tr("Exit the application"));
+    connect(m_exitAction, &QAction::triggered, this, &MainWindow::onExit);
+
+    // Edit actions
+    m_undoAction = new QAction(tr("&Undo"), this);
+    m_undoAction->setShortcut(QKeySequence::Undo);
+    m_undoAction->setStatusTip(tr("Undo the last operation"));
+    connect(m_undoAction, &QAction::triggered, this, &MainWindow::onUndo);
+
+    m_redoAction = new QAction(tr("&Redo"), this);
+    m_redoAction->setShortcut(QKeySequence::Redo);
+    m_redoAction->setStatusTip(tr("Redo the last undone operation"));
+    connect(m_redoAction, &QAction::triggered, this, &MainWindow::onRedo);
+
+    m_cutAction = new QAction(tr("Cu&t"), this);
+    m_cutAction->setShortcut(QKeySequence::Cut);
+    m_cutAction->setStatusTip(tr("Cut the selection to clipboard"));
+    connect(m_cutAction, &QAction::triggered, this, &MainWindow::onCut);
+
+    m_copyAction = new QAction(tr("&Copy"), this);
+    m_copyAction->setShortcut(QKeySequence::Copy);
+    m_copyAction->setStatusTip(tr("Copy the selection to clipboard"));
+    connect(m_copyAction, &QAction::triggered, this, &MainWindow::onCopy);
+
+    m_pasteAction = new QAction(tr("&Paste"), this);
+    m_pasteAction->setShortcut(QKeySequence::Paste);
+    m_pasteAction->setStatusTip(tr("Paste from clipboard"));
+    connect(m_pasteAction, &QAction::triggered, this, &MainWindow::onPaste);
+
+    logger.debug("Actions created successfully");
+}
+
+void MainWindow::createMenus() {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Creating menus");
+
+    // File menu
+    m_fileMenu = menuBar()->addMenu(tr("&File"));
+    m_fileMenu->addAction(m_newAction);
+    m_fileMenu->addAction(m_openAction);
+    m_fileMenu->addSeparator();
+    m_fileMenu->addAction(m_saveAction);
+    m_fileMenu->addAction(m_saveAsAction);
+    m_fileMenu->addSeparator();
+    m_fileMenu->addAction(m_exitAction);
+
+    // Edit menu
+    m_editMenu = menuBar()->addMenu(tr("&Edit"));
+    m_editMenu->addAction(m_undoAction);
+    m_editMenu->addAction(m_redoAction);
+    m_editMenu->addSeparator();
+    m_editMenu->addAction(m_cutAction);
+    m_editMenu->addAction(m_copyAction);
+    m_editMenu->addAction(m_pasteAction);
+
+    logger.debug("Menus created successfully");
+}
+
+void MainWindow::createToolbars() {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Creating toolbars");
+
+    // File toolbar
+    m_fileToolbar = addToolBar(tr("File"));
+    m_fileToolbar->addAction(m_newAction);
+    m_fileToolbar->addAction(m_openAction);
+    m_fileToolbar->addAction(m_saveAction);
+
+    logger.debug("Toolbars created successfully");
+}
+
+void MainWindow::createStatusBar() {
+    auto& logger = core::Logger::getInstance();
+    logger.debug("Creating status bar");
+
+    statusBar()->showMessage(tr("Ready"), 3000);  // Show for 3 seconds
+
+    logger.debug("Status bar created successfully");
+}
+
+// Slots implementation
+void MainWindow::onNewDocument() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: New Document");
+    statusBar()->showMessage(tr("New document created"), 2000);
+}
+
+void MainWindow::onOpenDocument() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Open Document");
+    statusBar()->showMessage(tr("Open document dialog (not implemented)"), 2000);
+}
+
+void MainWindow::onSaveDocument() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Save Document");
+    statusBar()->showMessage(tr("Document saved (not implemented)"), 2000);
+}
+
+void MainWindow::onSaveAsDocument() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Save As Document");
+    statusBar()->showMessage(tr("Save As dialog (not implemented)"), 2000);
+}
+
+void MainWindow::onExit() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Exit");
+    QApplication::quit();
+}
+
+void MainWindow::onUndo() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Undo");
+    statusBar()->showMessage(tr("Undo (not implemented)"), 2000);
+}
+
+void MainWindow::onRedo() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Redo");
+    statusBar()->showMessage(tr("Redo (not implemented)"), 2000);
+}
+
+void MainWindow::onCut() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Cut");
+    statusBar()->showMessage(tr("Cut (not implemented)"), 2000);
+}
+
+void MainWindow::onCopy() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Copy");
+    statusBar()->showMessage(tr("Copy (not implemented)"), 2000);
+}
+
+void MainWindow::onPaste() {
+    auto& logger = core::Logger::getInstance();
+    logger.info("Action triggered: Paste");
+    statusBar()->showMessage(tr("Paste (not implemented)"), 2000);
+}
+
+} // namespace gui
+} // namespace kalahari